update test pt-10

+++ tests/abstract_broker_client_tests/integrate_queue.py
@@ -721,7 +721,7 @@ class PubSubQueue:
             try:
                 indexes_unacked = [
                     x
-                    for x in range(len(all_recvd) * (3 / 2))
+                    for x in range(int(len(all_recvd) * (3 / 2)))  # math.ceil?
                     if not x % 2 and not x % 3
                 ]
                 print(indexes_unacked)
